Constant offsets work, non working example for dx/ddx

+++ xtrack/aperture/headers/base.h
@@ -4,7 +4,7 @@
 #include <math.h>
 #include <stdlib.h>
 
-typedef float float_type;
+typedef double float_type;
 
 #ifndef M_PI
 #define M_PI 3.14159265358979323846
